Moved highlighter rect into it's own class
